requirements.txt
typer>=0.12.3
google-generativeai>=0.8.2
rich>=13.7.1

